Removed circual input temporarily
